pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smartdigest-tgbot"
version = "0.1.0"
description = "Headless Telegram bot that forwards posts from public channels and sends AI digests."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "apscheduler>=3.10.4,<4.0.0",
  "beautifulsoup4>=4.12.3,<5.0.0",
  "httpx>=0.27.0,<1.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "python-telegram-bot>=21.0,<22.0",
  "PyYAML>=6.0.1,<7.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.1,<9.0.0",
  "pytest-asyncio>=0.23.5,<1.0.0",
  "detect-secrets>=1.5.0,<2.0.0",
]

[project.scripts]
smartdigest-bot = "smartdigest_bot.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
